refactor: decompose operations.py into 11 domain modules
The 2325-line operations.py violated the atomic-design 500-line
limit and blocked further additions. Split by responsibility into:

  _branch_ops.py     branch creation, checkout, listing, merge-base
  _commit_ops.py     status, commit, log, show, blame
  _config_ops.py     config get/set/list + validators
  _diff_ops.py       diff variants + validators
  _init_ops.py       repo init
  _rebase_ops.py     rebase, merge, cherry-pick, abort, continue
  _remote_ops.py     push, pull, fetch, remote management
  _staging_ops.py    add, reset, file existence helpers
  _stash_ops.py      stash list/push/pop/drop
  _submodule_ops.py  submodule status/add/update/sync
  _tag_ops.py        tag list/create/delete

All new modules are under 500 lines (largest: _remote_ops at 351).
operations.py becomes a thin re-export shim (85 lines) preserving
the public API for tests and external callers.

Four pre-existing tests patched module-level imports at the old
location (operations.subprocess, operations.Path,
operations._validate_diff_parameters). Retargeted those patches to
the new submodules; function imports through the shim are unchanged.
