+++ src/mcp_server_git/git/operations.py
@@ -1,20 +1,71 @@
-"""Git operations for MCP Git Server"""
+"""Git operations for MCP Git Server (re-export shim).
 
-import logging
-import os
-import re
-import subprocess
-from pathlib import Path
+Decomposed for AG file-size policy. See _*_ops.py modules for implementations.
+Public API preserved: tests and `git/__init__.py` continue to import from here.
+"""
 
-# Safe git import that handles ClaudeCode redirector conflicts
-from ..utils.git_import import GitCommandError, Repo
-
-logger = logging.getLogger(__name__)
-
-# Constants for timeout and validation
-CLI_AUTH_TIMEOUT = 10  # seconds for GitHub CLI authentication
-PUSH_OPERATION_TIMEOUT = 300  # seconds (5 minutes) for push operations
-MIN_TOKEN_LENGTH = 10  # minimum length for a valid GitHub token
+from ._diff_ops import (
+    _apply_diff_size_limiting,
+    _validate_commit_range,
+    _validate_diff_parameters,
+    git_diff,
+    git_diff_branches,
+    git_diff_staged,
+    git_diff_unstaged,
+)
+from ._commit_ops import (
+    git_blame,
+    git_commit,
+    git_log,
+    git_show,
+    git_status,
+)
+from ._staging_ops import git_add, git_reset
+from ._branch_ops import (
+    git_branch_list,
+    git_checkout,
+    git_create_branch,
+    git_merge_base,
+)
+from ._init_ops import git_init
+from ._remote_ops import (
+    git_fetch,
+    git_pull,
+    git_push,
+    git_remote_add,
+    git_remote_get_url,
+    git_remote_list,
+    git_remote_remove,
+    git_remote_rename,
+    git_remote_set_url,
+)
+from ._rebase_ops import (
+    git_abort,
+    git_cherry_pick,
+    git_continue,
+    git_merge,
+    git_rebase,
+)
+from ._stash_ops import (
+    git_stash_drop,
+    git_stash_list,
+    git_stash_pop,
+    git_stash_push,
+)
+from ._tag_ops import git_tag_create, git_tag_delete, git_tag_list
+from ._submodule_ops import (
+    git_submodule_add,
+    git_submodule_status,
+    git_submodule_sync,
+    git_submodule_update,
+)
+from ._config_ops import (
+    _validate_config_file,
+    _validate_config_key,
+    git_config_get,
+    git_config_list,
+    git_config_set,
+)
 
 __all__ = [
     "git_status",
@@ -41,9 +92,17 @@ __all__ = [
     "git_remote_remove",
     "git_remote_list",
     "git_remote_get_url",
+    "git_remote_rename",
+    "git_remote_set_url",
     "git_diff_branches",
+    "git_stash_list",
     "git_stash_push",
     "git_stash_pop",
+    "git_stash_drop",
+    "git_tag_list",
+    "git_tag_create",
+    "git_tag_delete",
+    "git_blame",
     "git_branch_list",
     "git_merge_base",
     "git_submodule_status",
@@ -53,2273 +112,10 @@ __all__ = [
     "git_config_get",
     "git_config_set",
     "git_config_list",
+    # Private helpers re-exported for test compatibility
+    "_apply_diff_size_limiting",
+    "_validate_commit_range",
+    "_validate_diff_parameters",
+    "_validate_config_file",
+    "_validate_config_key",
 ]
-
-
-def _validate_commit_range(commit_range: str) -> tuple[bool, str]:
-    """Validate commit range format and return (is_valid, error_message)
-
-    Supported formats:
-    - commit1..commit2 (range between commits)
-    - commit1...commit2 (symmetric difference)
-    - HEAD~1..HEAD (relative references)
-    - branch1..branch2 (branch ranges)
-    """
-    if not commit_range or not commit_range.strip():
-        return False, "Commit range cannot be empty"
-
-    # Patterns that should pass without warnings
-    valid_patterns = [
-        # Git commit hashes (6+ hex characters)
-        r"^[a-fA-F0-9]{6,40}\.{2,3}[a-fA-F0-9]{6,40}$",
-        # HEAD references with optional tilde notation
-        r"^HEAD~?\d*\.{2,3}HEAD~?\d*$",
-        # Branch name patterns (common patterns only - conservative)
-        r"^[a-zA-Z][a-zA-Z0-9_]{1,}\.{2,3}[a-zA-Z][a-zA-Z0-9_]{1,}$",
-        r"^(main|master|develop|dev|production|prod|staging|stage|test|release-v\d+\.\d+)\.{2,3}[a-zA-Z][a-zA-Z0-9_\-\.]{1,}$",
-        r"^[a-zA-Z][a-zA-Z0-9_\-\.]{1,}\.{2,3}(main|master|develop|dev|production|prod|staging|stage|test)$",
-        # Feature/release branch patterns with slashes
-        r"^(feature|bugfix|hotfix|release)/[a-zA-Z0-9\-_\.]+\.{2,3}[a-zA-Z][a-zA-Z0-9_/\-\.]*$",
-        r"^[a-zA-Z][a-zA-Z0-9_/\-\.]*\.{2,3}(feature|bugfix|hotfix|release)/[a-zA-Z0-9\-_\.]+$",
-        # Mixed patterns - hash with branches (minimum reasonable lengths)
-        r"^[a-fA-F0-9]{6,40}\.{2,3}[a-zA-Z][a-zA-Z0-9_/\-\.]{1,}$",
-        r"^[a-zA-Z][a-zA-Z0-9_/\-\.]{1,}\.{2,3}[a-fA-F0-9]{6,40}$",
-        # HEAD with branches/hashes (minimum reasonable lengths)
-        r"^HEAD~?\d*\.{2,3}[a-zA-Z][a-zA-Z0-9_/\-\.]{1,}$",
-        r"^[a-zA-Z][a-zA-Z0-9_/\-\.]{1,}\.{2,3}HEAD~?\d*$",
-        r"^HEAD~?\d*\.{2,3}[a-fA-F0-9]{6,40}$",
-        r"^[a-fA-F0-9]{6,40}\.{2,3}HEAD~?\d*$",
-    ]
-
-    range_patterns = valid_patterns
-
-    commit_range = commit_range.strip()
-
-    # Check against known patterns
-    for pattern in range_patterns:
-        if re.match(pattern, commit_range):
-            return True, ""
-
-    # Check for obvious injection attempts
-    dangerous_chars = [";", "|", "&", "`", "$", "(", ")"]
-    if any(char in commit_range for char in dangerous_chars):
-        return False, f"Invalid characters detected in commit range: {commit_range}"
-
-    # If no pattern matches, it might still be valid (git is flexible)
-    # But warn about unusual format
-    return (
-        True,
-        f"Warning: Unusual commit range format '{commit_range}' - proceed with caution",
-    )
-
-
-def _validate_diff_parameters(
-    target: str | None = None,
-    commit_range: str | None = None,
-    base_commit: str | None = None,
-    target_commit: str | None = None,
-) -> tuple[bool, str]:
-    """Validate that diff parameters are not conflicting or ambiguous"""
-    provided_params = []
-    if target:
-        provided_params.append("target")
-    if commit_range:
-        provided_params.append("commit_range")
-    if base_commit and target_commit:
-        provided_params.append("base_commit + target_commit")
-    elif base_commit or target_commit:
-        return False, "Both base_commit and target_commit must be provided together"
-
-    if len(provided_params) > 1:
-        return (
-            False,
-            f"Conflicting diff parameters: {', '.join(provided_params)}. Use only one method to specify what to diff.",
-        )
-
-    # Validate commit_range format if provided
-    if commit_range:
-        is_valid, error_msg = _validate_commit_range(commit_range)
-        if not is_valid:
-            return False, f"Invalid commit_range: {error_msg}"
-        elif error_msg:  # Warning case
-            return True, error_msg
-
-    return True, ""
-
-
-def _apply_diff_size_limiting(
-    diff_output: str,
-    operation_name: str,
-    stat_only: bool = False,
-    max_lines: int | None = None,
-) -> str:
-    """Apply size limiting to diff outputs with consistent formatting"""
-    if not diff_output.strip():
-        return f"No changes detected in {operation_name}"
-
-    if stat_only:
-        # This should be handled by the caller using --stat flag
-        return diff_output
-
-    # Apply line limit if specified
-    if max_lines and max_lines > 0:
-        lines = diff_output.split("\n")
-        if len(lines) > max_lines:
-            truncated_output = "\n".join(lines[:max_lines])
-            truncated_output += (
-                f"\n\n... [Truncated: showing {max_lines} of {len(lines)} lines]"
-            )
-            truncated_output += "\nUse stat_only=true for summary or increase max_lines for more content"
-            return truncated_output
-
-    # Check if output is extremely large and warn
-    if len(diff_output) > 50000:  # 50KB threshold
-        lines_count = len(diff_output.split("\n"))
-        warning = f"⚠️  Large diff detected ({lines_count} lines, ~{len(diff_output) // 1000}KB)\n"
-        warning += "Consider using stat_only=true for summary or max_lines parameter to limit output\n\n"
-        return warning + diff_output
-
-    return diff_output
-
-
-def git_status(repo: Repo, porcelain: bool = False) -> str:
-    """Get repository status in either human-readable or machine-readable format.
-
-    Args:
-        repo: Git repository object
-        porcelain: If True, return porcelain (machine-readable) format
-
-    Returns:
-        Status output string
-    """
-    if porcelain:
-        return repo.git.status("--porcelain")
-    else:
-        return repo.git.status()
-
-
-def git_diff_unstaged(
-    repo: Repo,
-    stat_only: bool = False,
-    max_lines: int | None = None,
-    name_only: bool = False,
-    paths: list[str] | None = None,
-) -> str:
-    """Get unstaged changes diff with file-specific and advanced options"""
-    try:
-        # Build git diff arguments
-        diff_args = []
-
-        # Add options based on parameters
-        if name_only:
-            diff_args.append("--name-only")
-        elif stat_only:
-            diff_args.append("--stat")
-
-        # Add specific paths if provided
-        if paths:
-            diff_args.extend(["--"] + paths)
-
-        # Execute git diff with arguments
-        diff_output = repo.git.diff(*diff_args)
-
-        # Handle name-only output
-        if name_only:
-            return (
-                f"Files with unstaged changes:\n{diff_output}"
-                if diff_output.strip()
-                else "No unstaged changes"
-            )
-
-        # Handle stat-only output
-        if stat_only:
-            return (
-                f"Unstaged changes summary:\n{diff_output}"
-                if diff_output.strip()
-                else "No unstaged changes"
-            )
-
-        # Apply size limiting for full diff output
-        return _apply_diff_size_limiting(
-            diff_output, "unstaged changes", stat_only, max_lines
-        )
-
-    except GitCommandError as e:
-        return f"❌ Diff unstaged failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Diff unstaged error: {str(e)}"
-
-
-def git_diff_staged(
-    repo: Repo,
-    stat_only: bool = False,
-    max_lines: int | None = None,
-    name_only: bool = False,
-    paths: list[str] | None = None,
-) -> str:
-    """Get staged changes diff with file-specific and advanced options"""
-    try:
-        # Build git diff arguments
-        diff_args = ["--cached"]
-
-        # Add options based on parameters
-        if name_only:
-            diff_args.append("--name-only")
-        elif stat_only:
-            diff_args.append("--stat")
-
-        # Add specific paths if provided
-        if paths:
-            diff_args.extend(["--"] + paths)
-
-        # Execute git diff with arguments
-        diff_output = repo.git.diff(*diff_args)
-
-        # Handle name-only output
-        if name_only:
-            return (
-                f"Files with staged changes:\n{diff_output}"
-                if diff_output.strip()
-                else "No staged changes"
-            )
-
-        # Handle stat-only output
-        if stat_only:
-            return (
-                f"Staged changes summary:\n{diff_output}"
-                if diff_output.strip()
-                else "No staged changes"
-            )
-
-        # Apply size limiting for full diff output
-        return _apply_diff_size_limiting(
-            diff_output, "staged changes", stat_only, max_lines
-        )
-
-    except GitCommandError as e:
-        return f"❌ Diff staged failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Diff staged error: {str(e)}"
-
-
-def git_diff(
-    repo: Repo,
-    target: str | None = None,
-    stat_only: bool = False,
-    max_lines: int | None = None,
-    name_only: bool = False,
-    commit_range: str | None = None,
-    base_commit: str | None = None,
-    target_commit: str | None = None,
-    paths: list[str] | None = None,
-) -> str:
-    """Get diff with advanced options including commit ranges and file filtering.
-
-    Parameter Precedence (mutually exclusive - only one method should be used):
-    1. commit_range: Use git range syntax like "HEAD~1..HEAD" or "main..develop"
-    2. base_commit + target_commit: Compare two specific commits/branches
-    3. target: Compare working tree against specific branch/commit (default behavior)
-    4. None: Compare working tree against HEAD (fallback)
-
-    Args:
-        repo: Git repository object
-        target: Branch/commit to diff against (conflicts with commit_range or base_commit/target_commit)
-        stat_only: Show only file change statistics, not content
-        max_lines: Limit output to specified number of lines (overridden by name_only/stat_only)
-        name_only: Show only names of changed files
-        commit_range: Git range syntax like "HEAD~1..HEAD" (conflicts with other diff methods)
-        base_commit: Starting commit for comparison (requires target_commit)
-        target_commit: Ending commit for comparison (requires base_commit)
-        paths: Filter diff to specific files/directories
-
-    Returns:
-        Formatted diff output with validation warnings if applicable
-
-    Raises:
-        Returns error message if parameters are conflicting or invalid
-    """
-    # Validate parameters for conflicts and security
-    is_valid, validation_msg = _validate_diff_parameters(
-        target=target,
-        commit_range=commit_range,
-        base_commit=base_commit,
-        target_commit=target_commit,
-    )
-    if not is_valid:
-        return f"❌ Parameter validation failed: {validation_msg}"
-
-    # If there's a warning, include it in the output
-    validation_warning = validation_msg if validation_msg and is_valid else None
-    try:
-        # Build git diff arguments
-        diff_args = []
-
-        # Determine what we're diffing
-        diff_description = ""
-
-        if commit_range:
-            # Use commit range syntax like "HEAD~1..HEAD"
-            diff_args.append(commit_range)
-            diff_description = f"commit range {commit_range}"
-        elif base_commit and target_commit:
-            # Compare two specific commits
-            diff_args.extend([base_commit, target_commit])
-            diff_description = f"{base_commit}...{target_commit}"
-        elif target:
-            # Compare against target branch/commit (original behavior)
-            diff_args.append(target)
-            diff_description = f"against {target}"
-        else:
-            # Default to comparing working tree against HEAD
-            diff_args.append("HEAD")
-            diff_description = "against HEAD"
-
-        # Add options based on parameters
-        if name_only:
-            diff_args.append("--name-only")
-        elif stat_only:
-            diff_args.append("--stat")
-
-        # Add specific paths if provided
-        if paths:
-            diff_args.extend(["--"] + paths)
-
-        # Execute git diff with arguments
-        diff_output = repo.git.diff(*diff_args)
-
-        # Handle name-only output
-        if name_only:
-            result = (
-                f"Changed files {diff_description}:\n{diff_output}"
-                if diff_output.strip()
-                else f"No changes {diff_description}"
-            )
-            if validation_warning:
-                result = f"⚠️ {validation_warning}\n\n{result}"
-            return result
-
-        # Handle stat-only output
-        if stat_only:
-            result = (
-                f"Diff {diff_description} summary:\n{diff_output}"
-                if diff_output.strip()
-                else f"No differences {diff_description}"
-            )
-            if validation_warning:
-                result = f"⚠️ {validation_warning}\n\n{result}"
-            return result
-
-        # Apply size limiting for full diff output
-        result = _apply_diff_size_limiting(
-            diff_output, f"diff {diff_description}", stat_only, max_lines
-        )
-        if validation_warning:
-            result = f"⚠️ {validation_warning}\n\n{result}"
-        return result
-
-    except GitCommandError as e:
-        return f"❌ Diff failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Diff error: {str(e)}"
-
-
-def git_commit(
-    repo: Repo,
-    message: str,
-    amend: bool = False,  # If True, amend the most recent commit instead of creating new one
-    gpg_sign: bool = False,
-    gpg_key_id: str | None = None,
-    allow_empty: bool = False,
-) -> str:
-    """Commit staged changes with optional amend, GPG signing and automatic security enforcement"""
-    try:
-        # Import security functions locally to avoid circular imports
-        from .security import enforce_secure_git_config
-
-        # 🔒 SECURITY: Enforce secure configuration before committing
-        security_result = enforce_secure_git_config(repo, strict_mode=True)
-        security_messages = []
-        if "✅" in security_result:
-            security_messages.append("🔒 Security configuration enforced")
-
-        # Force GPG signing for all commits (SECURITY REQUIREMENT)
-        force_gpg = True
-
-        # Get GPG key from parameters, environment, or git config
-        if gpg_key_id:
-            force_key_id = gpg_key_id
-        else:
-            # Try environment variable first
-            env_key = os.getenv("GPG_SIGNING_KEY")
-            if env_key:
-                force_key_id = env_key
-            else:
-                # Fall back to git config
-                try:
-                    config_key = repo.config_reader().get_value("user", "signingkey")
-                    force_key_id = str(config_key)
-                except Exception:
-                    return "❌ Could not determine GPG signing key. Please configure GPG_SIGNING_KEY env var"
-
-        if force_gpg:
-            # Use git command directly for GPG signing
-            cmd = ["git", "commit"]
-            if amend:
-                cmd.append("--amend")
-            if allow_empty:
-                cmd.append("--allow-empty")
-            cmd.append(f"--gpg-sign={force_key_id}")
-            cmd.extend(["-m", message])
-
-            result = subprocess.run(
-                cmd, cwd=repo.working_dir, capture_output=True, text=True
-            )
-            if result.returncode == 0:
-                # Get the commit hash from git log
-                hash_result = subprocess.run(
-                    ["git", "rev-parse", "HEAD"],
-                    cwd=repo.working_dir,
-                    capture_output=True,
-                    text=True,
-                )
-                commit_hash = (
-                    hash_result.stdout.strip()[:8]
-                    if hash_result.returncode == 0
-                    else "unknown"
-                )
-
-                action = "amended" if amend else "created"
-                success_msg = (
-                    f"✅ Commit {commit_hash} {action} with VERIFIED GPG signature"
-                )
-                if security_messages:
-                    success_msg += f"\n{chr(10).join(security_messages)}"
-
-                # Add security reminder
-                success_msg += f"\n🔒 Enforced GPG signing with key {force_key_id}"
-                success_msg += (
-                    "\n⚠️  MCP Git Server used - no fallback to system git commands"
-                )
-
-                return success_msg
-            else:
-                return f"❌ Commit failed: {result.stderr}\n🔒 GPG signing was enforced but failed"
-        else:
-            # This path should never be reached due to force_gpg=True
-            return "❌ SECURITY VIOLATION: Unsigned commits are not allowed by MCP Git Server"
-
-    except GitCommandError as e:
-        return f"❌ Commit failed: {str(e)}\n🔒 Security enforcement may have prevented insecure operation"
-    except Exception as e:
-        return f"❌ Commit error: {str(e)}\n🔒 Verify repository security configuration"
-
-
-def git_add(
-    repo: Repo,
-    files: list[str] | None = None,
-    add_all: bool = False,
-    update_only: bool = False,
-    patterns: list[str] | None = None,
-) -> str:
-    """Add files to git staging area with support for batch operations and patterns
-
-    Args:
-        repo: Git repository object
-        files: List of specific file paths to add (traditional behavior)
-        add_all: If True, stage all changes including untracked files (equivalent to git add -A)
-        update_only: If True, stage only modifications and deletions, not new files (equivalent to git add -u)
-        patterns: List of glob patterns to match files (e.g., ["*.py", "src/**/*.js"])
-
-    Returns:
-        Success or error message string
-
-    Note:
-        Parameters are mutually exclusive:
-        - Use files for specific file paths
-        - Use add_all for staging all changes (git add -A)
-        - Use update_only for staging only tracked file changes (git add -u)
-        - Use patterns for glob-based file matching
-
-        When using patterns, overlapping patterns will stage files only once.
-        The count reflects unique files staged, not pattern matches.
-    """
-    try:
-        # Validate mutually exclusive parameters
-        provided_options = []
-        if files:
-            provided_options.append("files")
-        if add_all:
-            provided_options.append("add_all")
-        if update_only:
-            provided_options.append("update_only")
-        if patterns:
-            provided_options.append("patterns")
-
-        if len(provided_options) > 1:
-            return f"❌ Conflicting parameters: {', '.join(provided_options)}. Use only one method to specify what to add."
-
-        if len(provided_options) == 0:
-            return "❌ No files specified. Use files, add_all, update_only, or patterns parameter."
-
-        # Handle batch operations (add_all or update_only)
-        if add_all:
-            # Stage all changes including untracked files
-            repo.git.add("-A")
-            # Get count of staged changes
-            try:
-                staged_output = repo.git.diff("--cached", "--name-only")
-                staged_files = [
-                    f.strip() for f in staged_output.split("\n") if f.strip()
-                ]
-                count = len(staged_files)
-                return f"✅ Added {count} file(s) to staging area (all changes)"
-            except Exception:
-                return "✅ Added files to staging area (all changes)"
-
-        if update_only:
-            # Stage only modifications and deletions (no new files)
-            repo.git.add("-u")
-            # Get count of staged changes
-            try:
-                staged_output = repo.git.diff("--cached", "--name-only")
-                staged_files = [
-                    f.strip() for f in staged_output.split("\n") if f.strip()
-                ]
-                count = len(staged_files)
-                return f"✅ Added {count} file(s) to staging area (tracked updates)"
-            except Exception:
-                return "✅ Added files to staging area (tracked updates)"
-
-        # Handle pattern-based additions
-        if patterns:
-            # Validate patterns for safety (no command injection)
-            dangerous_chars = [";", "|", "&", "`", "$", "(", ")"]
-            for pattern in patterns:
-                if any(char in pattern for char in dangerous_chars):
-                    return f"❌ Invalid characters detected in pattern: {pattern}"
-
-            # Add files matching patterns
-            try:
-                repo.git.add(*patterns)
-                # Get list of what was added
-                staged_output = repo.git.diff("--cached", "--name-only")
-                staged_files = [
-                    f.strip() for f in staged_output.split("\n") if f.strip()
-                ]
-                if staged_files:
-                    return f"✅ Added {len(staged_files)} file(s) to staging area: {', '.join(patterns)}"
-                else:
-                    return f"⚠️ No files matched patterns: {', '.join(patterns)}"
-            except GitCommandError as e:
-                return f"❌ Pattern matching failed: {str(e)}"
-
-        # Traditional file-by-file behavior (backward compatible)
-        if files:
-            # Validate files exist or are known to git as changes
-            missing_files = []
-
-            # Get git status once and parse it for all files
-            status_output = repo.git.status("--porcelain")
-            status_files = set()
-            for status_line in status_output.split("\n"):
-                if status_line.strip() and len(status_line) >= 3:
-                    # Porcelain format: XY filename (where X is staged, Y is working tree)
-                    status_file = status_line[3:].strip()
-                    status_files.add(status_file)
-
-            # Check each file
-            for file in files:
-                if not _file_exists_or_in_git(repo, file, status_files):
-                    missing_files.append(file)
-
-            if missing_files:
-                return f"❌ Files not found: {', '.join(missing_files)}"
-
-            # Add files to staging area
-            repo.git.add(*files)
-
-            # Verify files were added
-            try:
-                # Use git diff --cached to get staged files (works in all cases)
-                staged_output = repo.git.diff("--cached", "--name-only")
-                staged_files = [
-                    f.strip() for f in staged_output.split("\n") if f.strip()
-                ]
-            except (GitCommandError, Exception):
-                # Fallback to traditional method
-                try:
-                    staged_files = [item.a_path for item in repo.index.diff("HEAD")]
-                except (GitCommandError, Exception):
-                    staged_files = []
-
-            added_files = [f for f in files if f in staged_files]
-
-            if added_files:
-                return f"✅ Added {len(added_files)} file(s) to staging area: {', '.join(added_files)}"
-            else:
-                return "⚠️ No changes detected in specified files"
-
-        # Fallback (should not be reached due to parameter validation above)
-        return "❌ No files specified. Use files, add_all, update_only, or patterns parameter."
-
-    except GitCommandError as e:
-        # Handle GitCommandError string representation variations
-        error_msg = str(e)
-        if "Git command failed" in error_msg:
-            return "❌ Git add failed: Git command failed"
-        else:
-            return f"❌ Git add failed: {error_msg}"
-    except Exception as e:
-        return f"❌ Git add failed: {str(e)}"
-
-
-def _file_exists_or_in_git(repo: Repo, file: str, status_files: set) -> bool:
-    """Check if a file exists on filesystem or is tracked in git status.
-
-    This helper handles both production and test environments with mocks.
-
-    Args:
-        repo: Git repository object
-        file: File path to check
-        status_files: Set of files from git status --porcelain
-
-    Returns:
-        True if file exists or is in git status, False otherwise
-    """
-    try:
-        repo_path = Path(repo.working_dir)
-
-        # Check if we're in a test environment with mocks
-        if hasattr(repo_path, "_mock_name") or str(type(repo_path).__name__) == "Mock":
-            # Test environment - try to work with mocked Path
-            file_path = _get_mocked_file_path(repo_path, repo.working_dir, file)
-            if hasattr(file_path, "exists") and callable(
-                getattr(file_path, "exists", None)
-            ):
-                file_exists: bool = bool(file_path.exists())  # type: ignore[union-attr]
-                if hasattr(file_path, "is_symlink") and callable(
-                    getattr(file_path, "is_symlink", None)
-                ):
-                    file_exists = file_exists or bool(file_path.is_symlink())  # type: ignore[union-attr]
-                return file_exists or file in status_files
-            # Mock doesn't have exists method - fall back to status_files
-            return file in status_files
-        else:
-            # Production - normal Path operations
-            file_path = repo_path / file
-            return file_path.exists() or file_path.is_symlink() or file in status_files
-    except Exception:
-        # If file check fails, rely on git status
-        return file in status_files
-
-
-def _get_mocked_file_path(repo_path, working_dir: str, file: str):
-    """Get file path object in test environment with mocks.
-
-    Handles various mocking scenarios for Path operations in tests.
-    """
-    try:
-        # Try the / operator
-        return repo_path / file
-    except (TypeError, AttributeError):
-        # Fallback: try to use Path class side_effect (when Path is mocked)
-        path_class = Path
-        side_effect = getattr(path_class, "side_effect", None)
-        if side_effect is not None and callable(side_effect):
-            import os
-
-            full_path = os.path.join(working_dir, file)
-            return side_effect(full_path)
-        else:
-            # Last resort: create a basic mock
-            from unittest.mock import Mock
-
-            file_path = Mock()
-            file_path.exists.return_value = "existing.py" in file
-            file_path.is_symlink.return_value = False
-            return file_path
-
-
-def git_reset(
-    repo: Repo,
-    mode: str | None = None,
-    target: str | None = None,
-    files: list[str] | None = None,
-) -> str:
-    """Reset repository with advanced options (--soft, --mixed, --hard)"""
-    try:
-        # Validate reset mode
-        valid_modes = ["soft", "mixed", "hard"]
-        if mode and mode not in valid_modes:
-            return (
-                f"❌ Invalid reset mode '{mode}'. Valid modes: {', '.join(valid_modes)}"
-            )
-
-        # Build git reset command
-        reset_args = []
-
-        # Add mode flag if specified
-        if mode:
-            reset_args.append(f"--{mode}")
-
-        # Add target if specified
-        if target:
-            # Validate target exists
-            try:
-                repo.git.rev_parse(target)
-            except GitCommandError:
-                return f"❌ Target '{target}' does not exist"
-            reset_args.append(target)
-
-        # Add files if specified
-        if files:
-            # Validate files exist
-            for file in files:
-                if not os.path.exists(os.path.join(repo.working_dir, file)):
-                    return f"❌ File '{file}' does not exist"
-            reset_args.extend(files)
-
-        # Special handling for file-specific reset
-        if files and not mode and not target:
-            # Default to mixed reset for files
-            reset_args.insert(0, "HEAD")
-
-        # Get status before reset for informative message
-        status_before = ""
-        if mode in ["mixed", "hard"] or not mode:
-            try:
-                staged_files = [
-                    item.a_path for item in repo.index.diff("HEAD") if item.a_path
-                ]
-                if staged_files:
-                    status_before = f"staged files: {', '.join(staged_files[:5])}"
-                    if len(staged_files) > 5:
-                        status_before += f" (and {len(staged_files) - 5} more)"
-            except Exception:
-                pass
-
-        if mode == "hard":
-            try:
-                modified_files = [
-                    item.a_path for item in repo.index.diff(None) if item.a_path
-                ]
-                if modified_files:
-                    mod_status = f"modified files: {', '.join(modified_files[:5])}"
-                    if len(modified_files) > 5:
-                        mod_status += f" (and {len(modified_files) - 5} more)"
-                    status_before = (
-                        f"{status_before}, {mod_status}"
-                        if status_before
-                        else mod_status
-                    )
-            except Exception:
-                pass
-
-        # Execute reset
-        if reset_args:
-            repo.git.reset(*reset_args)
-        else:
-            repo.git.reset()
-
-        # Build success message
-        if files:
-            return f"✅ Reset {len(files)} file(s): {', '.join(files)}"
-        elif mode == "soft":
-            return f"✅ Soft reset to {target if target else 'HEAD'} - keeping changes in index"
-        elif mode == "mixed" or not mode:
-            target_msg = f" to {target}" if target else ""
-            return f"✅ Mixed reset{target_msg} - {status_before if status_before else 'no staged changes'}"
-        elif mode == "hard":
-            target_msg = f" to {target}" if target else ""
-            return f"✅ Hard reset{target_msg} - {status_before if status_before else 'no changes'} discarded"
-        else:
-            # Fallback return (should not reach here)
-            return "✅ Reset completed"
-
-    except GitCommandError as e:
-        return f"❌ Reset failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Reset error: {str(e)}"
-
-
-def git_log(
-    repo: Repo,
-    max_count: int = 10,
-    oneline: bool = False,
-    graph: bool = False,
-    format_str: str | None = None,  # Renamed from 'format'
-    since: str | None = None,
-    until: str | None = None,
-    author: str | None = None,
-    grep: str | None = None,
-    files: list[str] | None = None,
-    branch: str | None = None,
-    reverse: bool = False,
-    merges: bool | None = None,
-) -> str:
-    """Get commit history with advanced filtering and formatting options
-
-    Args:
-        repo: Git repository object
-        max_count: Maximum number of commits to show (default: 10)
-        oneline: Compact "hash message" format (equivalent to --oneline)
-        graph: Show merge graph (--graph)
-        format_str: Custom format string (e.g., "%h - %s (%an)")
-        since: Date filter - commits after this date (e.g., "2024-01-01", "1 week ago")
-        until: Date filter - commits before this date (e.g., "yesterday", "2024-12-31")
-        author: Filter by commit author (email or name)
-        grep: Search commit messages (regex pattern)
-        files: Commits affecting specific files (list of file paths)
-        branch: Specific branch to show log for (default: current branch)
-        reverse: Reverse chronological order (oldest first)
-        merges: Filter merge commits (None=all, True=only merges, False=no merges)
-
-    Returns:
-        Formatted commit log output
-    """
-    try:
-        args = []
-
-        # Add branch if specified
-        if branch:
-            args.append(branch)
-
-        # Add count limit
-        if max_count is not None and max_count > 0:
-            args.extend(["-n", str(max_count)])
-
-        # Add formatting options
-        if oneline:
-            args.append("--oneline")
-        elif format_str:  # Use format_str
-            args.extend(["--pretty=format:" + format_str])
-
-        if graph:
-            args.append("--graph")
-
-        # Add date filters
-        if since:
-            args.extend(["--since", since])
-        if until:
-            args.extend(["--until", until])
-
-        # Add author filter
-        if author:
-            args.extend(["--author", author])
-
-        # Add message search
-        if grep:
-            args.extend(["--grep", grep])
-
-        # Add merge commit filter
-        if merges is True:
-            args.append("--merges")
-        elif merges is False:
-            args.append("--no-merges")
-
-        # Add reverse order
-        if reverse:
-            args.append("--reverse")
-
-        # Add file paths if specified (must come after --)
-        if files:
-            args.append("--")
-            args.extend(files)
-
-        # Get commit log
-        log_output = repo.git.log(*args)
-
-        if not log_output.strip():
-            return "No commits found matching the specified criteria"
-
-        return log_output
-
-    except GitCommandError as e:
-        return f"❌ Log failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Log error: {str(e)}"
-
-
-def git_create_branch(
-    repo: Repo,
-    branch_name: str,
-    base_branch: str | None = None,
-    start_point: str | None = None,
-) -> str:
-    """Create new branch from base - UPDATED VERSION 2024"""
-    try:
-        # INTEGRATED DEBUG LOGGING
-        # Use start_point if provided, otherwise fall back to base_branch
-        effective_base = start_point if start_point is not None else base_branch
-
-        # Check if branch already exists
-        existing_branches = [branch.name for branch in repo.branches]
-        if branch_name in existing_branches:
-            return f"❌ Branch '{branch_name}' already exists"
-
-        # Create new branch
-        if effective_base:
-            # Build list of all valid branch references
-            all_branches = list(existing_branches)
-
-            # Add remote branches if remotes exist
-            try:
-                if repo.remotes:
-                    for remote in repo.remotes:
-                        # Include both full remote ref names (e.g., 'origin/main')
-                        # and short names (e.g., 'main') for compatibility
-                        all_branches.extend([ref.name for ref in remote.refs])
-                        all_branches.extend(
-                            [ref.name.split("/")[-1] for ref in remote.refs]
-                        )
-            except Exception:
-                # Ignore remote access errors (e.g., no remotes configured)
-                pass
-
-            # Verify base branch exists
-            if effective_base not in all_branches:
-                return f"❌ Base branch '{effective_base}' not found"
-
-            repo.create_head(branch_name, effective_base)
-        else:
-            repo.create_head(branch_name)
-
-        return f"✅ Created branch '{branch_name}'"
-
-    except GitCommandError as e:
-        return f"❌ Branch creation failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Branch creation error: {str(e)}"
-
-
-def git_checkout(repo: Repo, branch_name: str) -> str:
-    """Switch to a branch"""
-    try:
-        # Check if branch exists locally
-        local_branches = [branch.name for branch in repo.branches]
-
-        if branch_name in local_branches:
-            # Switch to local branch
-            repo.git.checkout(branch_name)
-            return f"✅ Switched to branch '{branch_name}'"
-        else:
-            # Check if it's a full remote ref (e.g., 'origin/development')
-            try:
-                remote_refs = [ref.name for ref in repo.remote().refs]
-                if branch_name in remote_refs:
-                    # Checkout the remote ref directly (detached HEAD)
-                    repo.git.checkout(branch_name)
-                    return f"✅ Switched to '{branch_name}' (detached HEAD)"
-            except Exception:
-                pass
-
-            # Check if branch exists on remote (short name)
-            try:
-                remote_branches = [
-                    ref.name.split("/")[-1] for ref in repo.remote().refs
-                ]
-                if branch_name in remote_branches:
-                    # Create local tracking branch
-                    repo.git.checkout("-b", branch_name, f"origin/{branch_name}")
-                    return f"✅ Created and switched to branch '{branch_name}' (tracking origin/{branch_name})"
-                else:
-                    return f"❌ Branch '{branch_name}' not found locally or on remote"
-            except Exception:
-                return f"❌ Branch '{branch_name}' not found"
-
-    except GitCommandError as e:
-        return f"❌ Checkout failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Checkout error: {str(e)}"
-
-
-def git_show(
-    repo: Repo, revision: str, stat_only: bool = False, max_lines: int | None = None
-) -> str:
-    """Show commit details with diff and size limiting options"""
-    try:
-        if stat_only:
-            # Return only commit info and file statistics
-            show_output = repo.git.show("--stat", revision)
-            return f"Commit details for {revision}:\n{show_output}"
-
-        # Get full commit details
-        show_output = repo.git.show(revision)
-
-        # Apply line limit if specified
-        if max_lines and max_lines > 0:
-            lines = show_output.split("\n")
-            if len(lines) > max_lines:
-                truncated_output = "\n".join(lines[:max_lines])
-                truncated_output += (
-                    f"\n\n... [Truncated: showing {max_lines} of {len(lines)} lines]"
-                )
-                truncated_output += "\nUse stat_only=true for summary or increase max_lines for more content"
-                return truncated_output
-
-        # Check if output is extremely large and warn
-        if len(show_output) > 50000:  # 50KB threshold
-            lines_count = len(show_output.split("\n"))
-            warning = f"⚠️  Large commit detected ({lines_count} lines, ~{len(show_output) // 1000}KB)\n"
-            warning += "Consider using stat_only=true for summary or max_lines parameter to limit output\n\n"
-            return warning + show_output
-
-        return show_output
-
-    except GitCommandError as e:
-        return f"❌ Show failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Show error: {str(e)}"
-
-
-def git_init(repo_path: str) -> str:
-    """Initialize new Git repository"""
-    try:
-        path = Path(repo_path)
-        path.mkdir(parents=True, exist_ok=True)
-
-        # Initialize repository
-        Repo.init(path)
-
-        return f"✅ Initialized empty Git repository in {repo_path}"
-
-    except Exception as e:
-        return f"❌ Init failed: {str(e)}"
-
-
-def _get_github_token_from_cli() -> str | None:
-    """Extract token from GitHub CLI if available"""
-    try:
-        logger.debug("🔍 DEBUG: Running 'gh auth token' command...")
-        result = subprocess.run(
-            ["gh", "auth", "token"],
-            capture_output=True,
-            text=True,
-            timeout=CLI_AUTH_TIMEOUT,
-        )
-        logger.debug(f"🔍 DEBUG: gh auth token return code: {result.returncode}")
-        logger.debug(
-            f"🔍 DEBUG: gh auth token stdout: {result.stdout[:50]}..."
-            if result.stdout
-            else "🔍 DEBUG: gh auth token stdout: EMPTY"
-        )
-        logger.debug(f"🔍 DEBUG: gh auth token stderr: {result.stderr}")
-
-        if result.returncode == 0:
-            token = result.stdout.strip()
-            token_valid = token and len(token) >= MIN_TOKEN_LENGTH
-            logger.debug(
-                f"🔍 DEBUG: Token valid: {token_valid}, length: {len(token) if token else 0}"
-            )
-            return token if token_valid else None
-    except (FileNotFoundError, subprocess.TimeoutExpired) as e:
-        logger.debug(f"🔍 DEBUG: gh command failed: {e}")
-        pass
-    return None
-
-
-def git_push(
-    repo: Repo,
-    remote: str = "origin",
-    branch: str | None = None,
-    set_upstream: bool = False,
-    force: bool = False,
-) -> str:
-    """Push with comprehensive authentication including fallback to system git credentials"""
-    try:
-        # Get current branch if not specified
-        if not branch:
-            try:
-                branch = repo.active_branch.name
-            except TypeError:  # Detached HEAD or no commits
-                return "❌ No active branch found and no branch specified"
-
-        # Build push arguments
-        push_args = [remote]
-        if branch:
-            push_args.append(branch)
-
-        if set_upstream:
-            push_args.insert(0, "--set-upstream")
-        if force:
-            push_args.insert(0, "--force")
-
-        # Get remote URL for GitHub authentication handling (cache for reuse)
-        remote_url = ""
-        is_github = False
-        try:
-            remote_url = repo.remote(remote).url
-            is_github = "github.com" in remote_url
-        except Exception:
-            pass
-
-        # GitHub HTTPS authentication handling
-        if is_github and remote_url.startswith("https://"):
-            # Try to load .env from current repository first
-            from dotenv import load_dotenv
-
-            repo_env = Path(repo.working_dir) / ".env"
-            if repo_env.exists():
-                logger.info(f"🔍 DEBUG: Loading .env from repository: {repo_env}")
-                load_dotenv(repo_env, override=True)
-
-            github_token = os.getenv("GITHUB_TOKEN")
-            logger.info(
-                f"🔍 DEBUG: GITHUB_TOKEN from env: {'SET' if github_token else 'NOT SET'}"
-            )
-            logger.info(f"🔍 DEBUG: Repository working dir: {repo.working_dir}")
-            logger.info(f"🔍 DEBUG: .env file exists: {repo_env.exists()}")
-
-            # If no GITHUB_TOKEN, try to get token from GitHub CLI
-            if not github_token:
-                logger.debug("🔍 DEBUG: Attempting GitHub CLI token extraction...")
-                github_token = _get_github_token_from_cli()
-                logger.debug(
-                    f"🔍 DEBUG: GitHub CLI token: {'SET' if github_token else 'NOT SET'}"
-                )
-
-            if github_token:
-                logger.debug(
-                    "🔍 DEBUG: Token found, proceeding with authenticated push"
-                )
-                # Inject token into URL
-                if "github.com" in remote_url:
-                    # Format: https://token@github.com/user/repo.git
-                    auth_url = remote_url.replace(
-                        "https://", f"https://{github_token}@"
-                    )
-
-                    # Temporarily set remote URL with token
-                    repo.remote(remote).set_url(auth_url)
-
-                    try:
-                        # Attempt push with authenticated URL
-                        repo.git.push(*push_args)
-                        success_msg = f"✅ Successfully pushed {branch} to {remote}"
-                        if set_upstream:
-                            success_msg += " (set upstream tracking)"
-
-                        # Indicate which authentication method was used
-                        if os.getenv("GITHUB_TOKEN"):
-                            success_msg += "\n🔐 Used GITHUB_TOKEN authentication"
-                        else:
-                            success_msg += "\n🔐 Used GitHub CLI authentication"
-                        return success_msg
-                    finally:
-                        # Restore original URL
-                        repo.remote(remote).set_url(remote_url)
-            else:
-                # Fallback to system git with credential helpers
-                logger.debug("🔍 DEBUG: NO TOKEN FOUND - falling back to system git")
-                logger.info(
-                    "No GitHub token available, falling back to system git authentication"
-                )
-                try:
-                    # Use subprocess to call system git with credential helpers
-                    cmd = ["git", "push"]
-                    cmd.extend(push_args)
-                    logger.debug(f"🔍 DEBUG: System git command: {' '.join(cmd)}")
-                    logger.debug(f"🔍 DEBUG: Working directory: {repo.working_dir}")
-
-                    result = subprocess.run(
-                        cmd,
-                        cwd=repo.working_dir,
-                        capture_output=True,
-                        text=True,
-                        timeout=PUSH_OPERATION_TIMEOUT,
-                    )
-
-                    logger.debug(
-                        f"🔍 DEBUG: System git return code: {result.returncode}"
-                    )
-                    logger.debug(f"🔍 DEBUG: System git stdout: {result.stdout}")
-                    logger.debug(f"🔍 DEBUG: System git stderr: {result.stderr}")
-
-                    if result.returncode == 0:
-                        success_msg = f"✅ Successfully pushed {branch} to {remote}"
-                        if set_upstream:
-                            success_msg += " (set upstream tracking)"
-                        success_msg += "\n🔐 Used system git authentication"
-                        return success_msg
-                    else:
-                        error_output = result.stderr.strip()
-                        if (
-                            "Authentication failed" in error_output
-                            or "401" in error_output
-                        ):
-                            # Add debug info directly to error message
-                            repo_env = Path(repo.working_dir) / ".env"
-                            token_status = (
-                                "SET" if os.getenv("GITHUB_TOKEN") else "NOT SET"
-                            )
-                            return (
-                                f"❌ Authentication failed. Configure GITHUB_TOKEN environment variable "
-                                f"or GitHub CLI authentication (gh auth login)\n"
-                                f"🔍 DEBUG: GITHUB_TOKEN: {token_status}, "
-                                f".env exists: {repo_env.exists()}, "
-                                f"working_dir: {repo.working_dir}\n"
-                                f"🔍 System git error: {error_output}"
-                            )
-                        elif (
-                            "403" in error_output or "Permission denied" in error_output
-                        ):
-                            return "❌ Permission denied. Check repository access permissions"
-                        elif "non-fast-forward" in error_output:
-                            return "❌ Push rejected (non-fast-forward). Use force=True if needed"
-                        else:
-                            return f"❌ Push failed: {error_output}"
-
-                except subprocess.TimeoutExpired:
-                    return "❌ Push operation timed out. Check network connection and repository access"
-                except Exception as e:
-                    return f"❌ System git push failed: {str(e)}"
-
-        # Regular push (SSH or authenticated HTTPS)
-        try:
-            repo.git.push(*push_args)
-            success_msg = f"✅ Successfully pushed {branch} to {remote}"
-            if set_upstream:
-                success_msg += " (set upstream tracking)"
-            return success_msg
-        except GitCommandError as e:
-            # If regular push fails and this is GitHub HTTPS, suggest auth options
-            if is_github and remote_url.startswith("https://"):
-                if "Authentication failed" in str(e) or "401" in str(e):
-                    # Add debug info directly to error message - REGULAR PUSH PATH
-                    repo_env = Path(repo.working_dir) / ".env"
-                    token_status = "SET" if os.getenv("GITHUB_TOKEN") else "NOT SET"
-                    return (
-                        f"❌ Authentication failed [DEBUG_VERSION_V3]. Configure GITHUB_TOKEN environment variable "
-                        f"or GitHub CLI authentication (gh auth login)\n"
-                        f"🔍 DEBUG [REGULAR_PUSH]: GITHUB_TOKEN: {token_status}, "
-                        f".env exists: {repo_env.exists()}, "
-                        f"working_dir: {repo.working_dir}\n"
-                        f"🔍 GitPython error: {str(e)}"
-                    )
-                elif "403" in str(e) or "Permission denied" in str(e):
-                    return "❌ Permission denied. Check repository access permissions"
-
-            # Standard error handling for non-GitHub or non-auth issues
-            if "non-fast-forward" in str(e):
-                return "❌ Push rejected (non-fast-forward). Use force=True if needed"
-            else:
-                return f"❌ Push failed: {str(e)}"
-
-    except GitCommandError as e:
-        if "Authentication failed" in str(e) or "401" in str(e):
-            # Add debug info directly to error message - OUTER EXCEPTION PATH
-            repo_env = Path(repo.working_dir) / ".env"
-            token = os.getenv("GITHUB_TOKEN", "")
-            token_info = (
-                f"length={len(token)}, starts_with={token[:4]}..."
-                if token
-                else "NOT SET"
-            )
-            return (
-                f"❌ Authentication failed. Configure GITHUB_TOKEN environment variable "
-                f"or GitHub CLI authentication (gh auth login)\n"
-                f"🔍 DEBUG [OUTER_EXCEPTION]: GITHUB_TOKEN: {token_info}, "
-                f".env exists: {repo_env.exists()}, "
-                f"working_dir: {repo.working_dir}\n"
-                f"🔍 Outer GitPython error: {str(e)}"
-            )
-        elif "403" in str(e):
-            return "❌ Permission denied. Check repository access permissions"
-        elif "non-fast-forward" in str(e):
-            return "❌ Push rejected (non-fast-forward). Use force=True if needed"
-        else:
-            return f"❌ Push failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Push error: {str(e)}"
-
-
-def git_pull(repo: Repo, remote: str = "origin", branch: str | None = None) -> str:
-    """Pull changes from remote repository"""
-    try:
-        # Get current branch if not specified
-        if not branch:
-            try:
-                branch = repo.active_branch.name
-            except TypeError:  # Detached HEAD or no commits
-                return "❌ No active branch found and no branch specified"
-
-        # Perform pull
-        if branch:
-            result = repo.git.pull(remote, branch)
-        else:
-            result = repo.git.pull(remote)
-
-        return f"✅ Successfully pulled from {remote}/{branch}\n{result}"
-
-    except GitCommandError as e:
-        if "Authentication failed" in str(e):
-            return f"❌ Authentication failed. Check credentials for {remote}"
-        elif "merge conflict" in str(e).lower():
-            return "❌ Pull failed due to merge conflicts. Resolve conflicts and retry"
-        else:
-            return f"❌ Pull failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Pull error: {str(e)}"
-
-
-def git_diff_branches(
-    repo: Repo,
-    base_branch: str,
-    compare_branch: str,
-    stat_only: bool = False,
-    max_lines: int | None = None,
-) -> str:
-    """Show differences between two branches with size limiting options"""
-    try:
-        # Verify branches exist - support both short names and full remote refs
-        # Special refs like HEAD are always valid
-        special_refs = ["HEAD", "FETCH_HEAD", "ORIG_HEAD", "MERGE_HEAD"]
-
-        local_branches = [branch.name for branch in repo.branches]
-
-        # Add remote branches if remotes exist (with error handling)
-        try:
-            remote_refs = repo.remote().refs
-            # Include both full remote ref names (e.g., 'origin/development')
-            # and short names (e.g., 'development') for compatibility
-            remote_branch_names = [ref.name for ref in remote_refs]
-            remote_branch_short_names = [ref.name.split("/")[-1] for ref in remote_refs]
-            # Use set to avoid duplicates
-            all_branches = set(
-                local_branches
-                + remote_branch_names
-                + remote_branch_short_names
-                + special_refs
-            )
-        except Exception:
-            # Ignore remote access errors (e.g., no remotes configured)
-            all_branches = set(local_branches + special_refs)
-
-        if base_branch not in all_branches:
-            return f"❌ Base branch '{base_branch}' not found"
-        if compare_branch not in all_branches:
-            return f"❌ Compare branch '{compare_branch}' not found"
-
-        # Build diff command arguments
-        diff_range = f"{base_branch}...{compare_branch}"
-
-        if stat_only:
-            # Return only file statistics
-            diff_output = repo.git.diff("--stat", diff_range)
-            if not diff_output.strip():
-                return f"No differences between {base_branch} and {compare_branch}"
-            return f"Diff statistics between {base_branch} and {compare_branch}:\n{diff_output}"
-
-        # Get full diff
-        diff_output = repo.git.diff(diff_range)
-
-        if not diff_output.strip():
-            return f"No differences between {base_branch} and {compare_branch}"
-
-        # Apply line limit if specified
-        if max_lines and max_lines > 0:
-            lines = diff_output.split("\n")
-            if len(lines) > max_lines:
-                truncated_output = "\n".join(lines[:max_lines])
-                truncated_output += (
-                    f"\n\n... [Truncated: showing {max_lines} of {len(lines)} lines]"
-                )
-                truncated_output += "\nUse --stat flag for summary or increase max_lines for more content"
-                return truncated_output
-
-        # Check if output is extremely large and warn
-        if len(diff_output) > 50000:  # 50KB threshold
-            lines_count = len(diff_output.split("\n"))
-            warning = f"⚠️  Large diff detected ({lines_count} lines, ~{len(diff_output) // 1000}KB)\n"
-            warning += "Consider using stat_only=true for summary or max_lines parameter to limit output\n\n"
-            return warning + diff_output
-
-        return diff_output
-
-    except GitCommandError as e:
-        return f"❌ Diff failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Diff error: {str(e)}"
-
-
-def git_rebase(
-    repo: Repo,
-    target_branch: str,
-    onto: str | None = None,
-    fork_point: str | None = None,
-    branch: str | None = None,
-) -> str:
-    """Rebase current branch onto target branch.
-
-    Supports --onto for rebasing between bases:
-        git rebase --onto <new_base> <old_fork_point> [<branch>]
-
-    Examples:
-        git_rebase(repo, "main")                          # simple rebase
-        git_rebase(repo, "main", branch="feature")        # rebase feature onto main
-        git_rebase(repo, "main",                          # --onto rebase
-                   onto="new-base", fork_point="old-base")
-    """
-    try:
-        # Validate ref parameters for shell injection
-        dangerous_chars = [";", "|", "&", "`", "$", "(", ")"]
-        for param_name, param_value in [
-            ("onto", onto),
-            ("fork_point", fork_point),
-            ("branch", branch),
-        ]:
-            if param_value and any(char in param_value for char in dangerous_chars):
-                return f"❌ Invalid characters detected in {param_name}: '{param_value}'"
-
-        # Validate onto/fork_point pairing
-        if onto and not fork_point:
-            return "❌ --onto requires fork_point (the old base to rebase from)"
-        if fork_point and not onto:
-            return "❌ fork_point requires --onto (the new base to rebase onto)"
-
-        # Get current branch for success message
-        current_branch = repo.active_branch.name
-
-        # Check if target branch exists (only when not using --onto)
-        if not onto:
-            all_branches = [b.name for b in repo.branches]
-            try:
-                if repo.remotes:
-                    for remote in repo.remotes:
-                        all_branches.extend([ref.name for ref in remote.refs])
-                        all_branches.extend(
-                            [ref.name.split("/")[-1] for ref in remote.refs]
-                        )
-            except Exception:
-                pass
-            if target_branch not in all_branches:
-                return f"❌ Target branch '{target_branch}' not found"
-
-        # Build rebase command args
-        if onto:
-            args = ["--onto", onto, fork_point]
-            if branch:
-                args.append(branch)
-        elif branch:
-            args = [target_branch, branch]
-        else:
-            args = [target_branch]
-
-        result = repo.git.rebase(*args)
-
-        rebase_branch = branch or current_branch
-        if onto:
-            return f"✅ Successfully rebased {rebase_branch} --onto {onto} (from {fork_point})\n{result}"
-        return f"✅ Successfully rebased {rebase_branch} onto {target_branch}\n{result}"
-
-    except GitCommandError as e:
-        if "conflict" in str(e).lower():
-            return "❌ Rebase failed due to conflicts. Resolve conflicts and run 'git rebase --continue'"
-        else:
-            return f"❌ Rebase failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Rebase error: {str(e)}"
-
-
-def git_merge(
-    repo: Repo,
-    source_branch: str,
-    strategy: str = "merge",
-    message: str | None = None,
-) -> str:
-    """Merge source branch with strategy options"""
-    try:
-        # Get current branch
-        current_branch = repo.active_branch.name
-
-        # Check if source branch exists - support both short names and full remote refs
-        all_branches = [branch.name for branch in repo.branches]
-
-        # Add remote branches if remotes exist
-        try:
-            if repo.remotes:
-                for remote in repo.remotes:
-                    # Include both full remote ref names (e.g., 'origin/development')
-                    # and short names (e.g., 'development') for compatibility
-                    all_branches.extend([ref.name for ref in remote.refs])
-                    all_branches.extend(
-                        [ref.name.split("/")[-1] for ref in remote.refs]
-                    )
-        except Exception:
-            # Ignore remote access errors (e.g., no remotes configured)
-            pass
-        if source_branch not in all_branches:
-            return f"❌ Source branch '{source_branch}' not found"
-
-        # Build merge command
-        merge_args = [source_branch]
-        if message:
-            merge_args.extend(["-m", message])
-
-        # Perform merge
-        result = repo.git.merge(*merge_args)
-
-        return f"✅ Successfully merged {source_branch} into {current_branch}\n{result}"
-
-    except GitCommandError as e:
-        if "conflict" in str(e).lower():
-            return "❌ Merge failed due to conflicts. Resolve conflicts and commit"
-        else:
-            return f"❌ Merge failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Merge error: {str(e)}"
-
-
-def git_cherry_pick(repo: Repo, commit_hash: str, no_commit: bool = False) -> str:
-    """Cherry-pick commits"""
-    try:
-        # Build cherry-pick command
-        cp_args = [commit_hash]
-        if no_commit:
-            cp_args.insert(0, "--no-commit")
-
-        # Perform cherry-pick
-        result = repo.git.cherry_pick(*cp_args)
-
-        action = "staged" if no_commit else "cherry-picked"
-        return f"✅ Successfully {action} commit {commit_hash[:8]}\n{result}"
-
-    except GitCommandError as e:
-        if "conflict" in str(e).lower():
-            return (
-                "❌ Cherry-pick failed due to conflicts. Resolve conflicts and continue"
-            )
-        else:
-            return f"❌ Cherry-pick failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Cherry-pick error: {str(e)}"
-
-
-def git_abort(repo: Repo, operation: str) -> str:
-    """Abort ongoing operations (rebase, merge, cherry-pick)"""
-    try:
-        valid_operations = ["rebase", "merge", "cherry-pick"]
-        if operation not in valid_operations:
-            return f"❌ Invalid operation '{operation}'. Valid operations: {', '.join(valid_operations)}"
-
-        # Perform abort using the same pattern as other operations
-        if operation == "rebase":
-            repo.git.rebase("--abort")
-        elif operation == "merge":
-            repo.git.merge("--abort")
-        elif operation == "cherry-pick":
-            repo.git.cherry_pick("--abort")
-
-        return f"✅ Successfully aborted {operation}"
-
-    except GitCommandError as e:
-        return f"❌ Abort {operation} failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Abort error: {str(e)}"
-
-
-def git_continue(repo: Repo, operation: str) -> str:
-    """Continue operations after resolving conflicts
-
-    Uses subprocess instead of GitPython for interactive operations to avoid
-    MCP client timeout issues (issue #97). GitPython's handling of interactive
-    git operations can trigger client-side AbortError (-32001).
-    """
-    try:
-        valid_operations = ["rebase", "merge", "cherry-pick"]
-        if operation not in valid_operations:
-            return f"❌ Invalid operation '{operation}'. Valid operations: {', '.join(valid_operations)}"
-
-        # Use subprocess for all continue operations to ensure reliable execution
-        # This avoids GitPython's interactive operation handling issues
-        cmd = ["git"]
-
-        if operation == "rebase":
-            cmd.extend(["rebase", "--continue"])
-        elif operation == "merge":
-            cmd.extend(["merge", "--continue"])
-        elif operation == "cherry-pick":
-            cmd.extend(["cherry-pick", "--continue"])
-
-        # Execute git command directly via subprocess
-        result = subprocess.run(
-            cmd,
-            cwd=repo.working_dir,
-            capture_output=True,
-            text=True,
-            timeout=60,  # 60 second timeout for continue operations
-        )
-
-        if result.returncode == 0:
-            # Success - combine stdout and stderr for complete output
-            output = (result.stdout + result.stderr).strip()
-            success_msg = f"✅ Successfully continued {operation}"
-            if output:
-                success_msg += f"\n{output}"
-            return success_msg
-        else:
-            # Failed - return stderr which contains the error message
-            error_output = result.stderr.strip()
-            if not error_output:
-                error_output = result.stdout.strip()
-
-            # Provide helpful error messages based on common scenarios
-            if (
-                "No rebase in progress" in error_output
-                or "no merge in progress" in error_output
-                or "no cherry-pick in progress" in error_output
-            ):
-                return f"❌ No {operation} in progress to continue"
-            elif "conflicts" in error_output.lower():
-                return f"❌ Unresolved conflicts remain. Resolve conflicts before continuing {operation}"
-            elif "nothing to commit" in error_output.lower():
-                return f"❌ No changes to commit. Add changes before continuing {operation}"
-            else:
-                return f"❌ Continue {operation} failed: {error_output}"
-
-    except subprocess.TimeoutExpired:
-        return f"❌ {operation} continue operation timed out after 60 seconds"
-    except GitCommandError as e:
-        return f"❌ Continue {operation} failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Continue error: {str(e)}"
-
-
-def git_remote_list(repo: Repo, verbose: bool = False) -> str:
-    """List all remote repositories"""
-    try:
-        if verbose:
-            return repo.git.remote("-v")
-        else:
-            return repo.git.remote()
-    except GitCommandError as e:
-        return f"❌ Remote list failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Remote list error: {str(e)}"
-
-
-def git_remote_add(repo: Repo, name: str, url: str) -> str:
-    """Add a new remote repository"""
-    try:
-        repo.git.remote("add", name, url)
-        return f"✅ Successfully added remote '{name}' -> {url}"
-    except GitCommandError as e:
-        return f"❌ Remote add failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Remote add error: {str(e)}"
-
-
-def git_remote_remove(repo: Repo, name: str) -> str:
-    """Remove a remote repository"""
-    try:
-        repo.git.remote("remove", name)
-        return f"✅ Successfully removed remote '{name}'"
-    except GitCommandError as e:
-        return f"❌ Remote remove failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Remote remove error: {str(e)}"
-
-
-def git_remote_rename(repo: Repo, old_name: str, new_name: str) -> str:
-    """Rename a remote repository"""
-    try:
-        repo.git.remote("rename", old_name, new_name)
-        return f"✅ Successfully renamed remote '{old_name}' to '{new_name}'"
-    except GitCommandError as e:
-        return f"❌ Remote rename failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Remote rename error: {str(e)}"
-
-
-def git_remote_set_url(repo: Repo, name: str, url: str) -> str:
-    """Set URL for a remote repository"""
-    try:
-        repo.git.remote("set-url", name, url)
-        return f"✅ Successfully set URL for remote '{name}' -> {url}"
-    except GitCommandError as e:
-        return f"❌ Remote set-url failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Remote set-url error: {str(e)}"
-
-
-def git_remote_get_url(repo: Repo, name: str) -> str:
-    """Get URL for a remote repository"""
-    try:
-        url = repo.git.remote("get-url", name)
-        return f"Remote '{name}' URL: {url}"
-    except GitCommandError as e:
-        return f"❌ Remote get-url failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Remote get-url error: {str(e)}"
-
-
-def git_fetch(
-    repo: Repo,
-    remote: str = "origin",
-    branch: str | None = None,
-    prune: bool = False,
-) -> str:
-    """Fetch changes from remote repository"""
-    try:
-        args = [remote]
-        if branch:
-            args.append(branch)
-        if prune:
-            args.append("--prune")
-
-        repo.git.fetch(*args)
-
-        if branch:
-            return f"✅ Successfully fetched {remote}/{branch}" + (
-                " (with prune)" if prune else ""
-            )
-        else:
-            return f"✅ Successfully fetched from {remote}" + (
-                " (with prune)" if prune else ""
-            )
-    except GitCommandError as e:
-        return f"❌ Fetch failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Fetch error: {str(e)}"
-
-
-def git_stash_list(repo: Repo) -> str:
-    """List all stashes"""
-    try:
-        stash_list = repo.git.stash("list")
-        if not stash_list.strip():
-            return "No stashes found"
-        return f"Stash list:\n{stash_list}"
-    except GitCommandError as e:
-        return f"❌ Stash list failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Stash list error: {str(e)}"
-
-
-def git_stash_push(
-    repo: Repo, message: str | None = None, include_untracked: bool = False
-) -> str:
-    """Create a new stash"""
-    try:
-        args = ["push"]
-        if include_untracked:
-            args.append("--include-untracked")
-        if message:
-            args.extend(["-m", message])
-
-        repo.git.stash(*args)
-        return "✅ Successfully created stash" + (f": {message}" if message else "")
-    except GitCommandError as e:
-        return f"❌ Stash push failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Stash push error: {str(e)}"
-
-
-def git_stash_pop(repo: Repo, stash_id: str | None = None) -> str:
-    """Apply and remove a stash"""
-    try:
-        if stash_id:
-            repo.git.stash("pop", stash_id)
-            return f"✅ Successfully popped stash {stash_id}"
-        else:
-            repo.git.stash("pop")
-            return "✅ Successfully popped latest stash"
-    except GitCommandError as e:
-        return f"❌ Stash pop failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Stash pop error: {str(e)}"
-
-
-def git_stash_drop(repo: Repo, stash_id: str | None = None) -> str:
-    """Remove a stash without applying it"""
-    try:
-        if stash_id:
-            repo.git.stash("drop", stash_id)
-            return f"✅ Successfully dropped stash {stash_id}"
-        else:
-            repo.git.stash("drop")
-            return "✅ Successfully dropped latest stash"
-    except GitCommandError as e:
-        return f"❌ Stash drop failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Stash drop error: {str(e)}"
-
-
-def git_tag_list(repo: Repo) -> str:
-    """List all tags"""
-    try:
-        tag_list = repo.git.tag("-l")
-        if not tag_list.strip():
-            return "No tags found"
-        return f"Tags:\n{tag_list}"
-    except GitCommandError as e:
-        return f"❌ Tag list failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Tag list error: {str(e)}"
-
-
-def git_tag_create(
-    repo: Repo,
-    tag_name: str,
-    message: str | None = None,
-    commit: str | None = None,
-) -> str:
-    """Create a new tag"""
-    try:
-        args = [tag_name]
-        if message:
-            args.extend(["-m", message])
-        if commit:
-            args.append(commit)
-
-        repo.git.tag(*args)
-        return f"✅ Successfully created tag '{tag_name}'" + (
-            f" on {commit}" if commit else ""
-        )
-    except GitCommandError as e:
-        return f"❌ Tag create failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Tag create error: {str(e)}"
-
-
-def git_tag_delete(repo: Repo, tag_name: str) -> str:
-    """Delete a tag"""
-    try:
-        repo.git.tag("-d", tag_name)
-        return f"✅ Successfully deleted tag '{tag_name}'"
-    except GitCommandError as e:
-        return f"❌ Tag delete failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Tag delete error: {str(e)}"
-
-
-def git_blame(
-    repo: Repo,
-    file_path: str,
-    line_start: int | None = None,
-    line_end: int | None = None,
-) -> str:
-    """Show blame information for a file"""
-    try:
-        args = [file_path]
-        if line_start and line_end:
-            args.extend(["-L", f"{line_start},{line_end}"])
-        elif line_start:
-            args.extend(["-L", f"{line_start},+1"])
-
-        blame_output = repo.git.blame(*args)
-        return f"Blame for {file_path}:\n{blame_output}"
-    except GitCommandError as e:
-        return f"❌ Blame failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Blame error: {str(e)}"
-
-
-def git_branch_list(
-    repo: Repo,
-    remote: bool = False,
-    all: bool = False,
-    pattern: str | None = None,
-) -> str:
-    """List branches in the repository
-
-    Args:
-        repo: Repository object
-        remote: If True, list remote branches (git branch -r)
-        all: If True, list all branches including remote (git branch -a)
-        pattern: Optional pattern to filter branches (supports glob patterns like 'feature/*')
-
-    Returns:
-        String containing the list of branches
-    """
-    try:
-        args = []
-
-        # Add branch listing flags
-        if all:
-            args.append("-a")
-        elif remote:
-            args.append("-r")
-
-        # When using pattern, we need to add --list flag
-        if pattern and pattern.strip():
-            args.append("--list")
-            args.append(pattern)
-
-        branch_output = repo.git.branch(*args)
-
-        if not branch_output.strip():
-            return "No branches found"
-
-        return f"Branches:\n{branch_output}"
-    except GitCommandError as e:
-        return f"❌ Branch list failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Branch list error: {str(e)}"
-
-
-def git_merge_base(
-    repo: Repo,
-    ref1: str,
-    ref2: str,
-) -> str:
-    """Find the common ancestor (merge-base) of two references.
-
-    Args:
-        repo: Git repository object
-        ref1: First reference (branch, commit, or tag)
-        ref2: Second reference (branch, commit, or tag)
-
-    Returns:
-        Formatted merge-base information including SHA, author, date, and message
-    """
-    try:
-        # Validate inputs for dangerous characters
-        dangerous_chars = [";", "|", "&", "`", "$", "(", ")"]
-        for ref_name, ref_value in [("ref1", ref1), ("ref2", ref2)]:
-            if any(char in ref_value for char in dangerous_chars):
-                return f"❌ Invalid characters detected in {ref_name}: {ref_value}"
-
-        # Execute git merge-base
-        merge_base_sha = repo.git.merge_base(ref1, ref2)
-
-        # Get commit details for context
-        commit = repo.commit(merge_base_sha)
-
-        # Format output
-        return (
-            f"Merge base: {merge_base_sha[:8]}\n"
-            f"Full SHA: {merge_base_sha}\n"
-            f"Author: {commit.author.name} <{commit.author.email}>\n"
-            f"Date: {commit.committed_datetime.isoformat()}\n"
-            f"Message: {commit.message.strip()}"
-        )
-    except GitCommandError as e:
-        return f"❌ Error finding merge-base: {str(e)}"
-    except Exception as e:
-        return f"❌ Merge-base error: {str(e)}"
-
-
-# ---------------------------------------------------------------------------
-# Submodule operations
-# ---------------------------------------------------------------------------
-
-
-def git_submodule_status(repo: Repo) -> str:
-    """List submodules and their current status.
-
-    Args:
-        repo: Git repository object
-
-    Returns:
-        Formatted list of submodules with paths, SHAs, and branches
-    """
-    try:
-        output = repo.git.submodule("status")
-        if not output.strip():
-            return "No submodules found in this repository"
-        lines = output.strip().split("\n")
-        result_lines = ["Submodules:"]
-        for line in lines:
-            line = line.strip()
-            if not line:
-                continue
-            # Parse status indicator
-            status = "up-to-date"
-            if line.startswith("+"):
-                status = "modified"
-                line = line[1:]
-            elif line.startswith("-"):
-                status = "not initialized"
-                line = line[1:]
-            elif line.startswith("U"):
-                status = "merge conflict"
-                line = line[1:]
-
-            parts = line.strip().split()
-            sha = parts[0] if parts else "unknown"
-            path = parts[1] if len(parts) > 1 else "unknown"
-            branch = parts[2].strip("()") if len(parts) > 2 else ""
-
-            branch_info = f" ({branch})" if branch else ""
-            result_lines.append(f"  {path}: {sha[:8]}{branch_info} [{status}]")
-
-        return "\n".join(result_lines)
-    except GitCommandError as e:
-        return f"❌ Submodule status failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Submodule status error: {str(e)}"
-
-
-def git_submodule_add(
-    repo: Repo,
-    url: str,
-    path: str,
-    branch: str | None = None,
-) -> str:
-    """Add a new submodule to the repository.
-
-    Args:
-        repo: Git repository object
-        url: URL of the submodule repository
-        path: Local path where the submodule will be placed
-        branch: Branch to track (optional)
-
-    Returns:
-        Success or error message
-    """
-    # Validate URL scheme
-    allowed_schemes = ("https://", "http://", "git://", "ssh://", "git@")
-    if not any(url.startswith(scheme) for scheme in allowed_schemes):
-        return f"❌ Invalid URL scheme. Allowed: {', '.join(allowed_schemes)}"
-    # Validate no shell injection in path or url
-    dangerous_chars = [";", "|", "&", "`", "$", "(", ")"]
-    if any(char in path for char in dangerous_chars):
-        return f"❌ Invalid characters in submodule path: {path}"
-    if any(char in url for char in dangerous_chars):
-        return f"❌ Invalid characters in URL: {url}"
-    try:
-        args = ["add"]
-        if branch:
-            args.extend(["--branch", branch])
-        args.extend([url, path])
-        repo.git.submodule(*args)
-        branch_info = f" (branch: {branch})" if branch else ""
-        return f"✅ Submodule added at '{path}' from {url}{branch_info}"
-    except GitCommandError as e:
-        return f"❌ Submodule add failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Submodule add error: {str(e)}"
-
-
-def git_submodule_update(
-    repo: Repo,
-    init: bool = True,
-    recursive: bool = False,
-    remote: bool = False,
-    paths: list[str] | None = None,
-) -> str:
-    """Update submodules (optionally init, recursive, or from remote).
-
-    Args:
-        repo: Git repository object
-        init: Initialize uninitialized submodules before updating
-        recursive: Recursively update nested submodules
-        remote: Update to latest remote commit instead of recorded SHA
-        paths: Specific submodule paths to update (default: all)
-
-    Returns:
-        Success or error message
-    """
-    try:
-        args = ["update"]
-        if init:
-            args.append("--init")
-        if recursive:
-            args.append("--recursive")
-        if remote:
-            args.append("--remote")
-        if paths:
-            args.extend(paths)
-        repo.git.submodule(*args)
-        scope = ", ".join(paths) if paths else "all submodules"
-        flags = []
-        if init:
-            flags.append("init")
-        if recursive:
-            flags.append("recursive")
-        if remote:
-            flags.append("remote")
-        flag_info = f" [{', '.join(flags)}]" if flags else ""
-        return f"✅ Submodules updated{flag_info}: {scope}"
-    except GitCommandError as e:
-        return f"❌ Submodule update failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Submodule update error: {str(e)}"
-
-
-def git_submodule_sync(
-    repo: Repo,
-    recursive: bool = False,
-) -> str:
-    """Sync submodule URLs from .gitmodules to .git/config.
-
-    Args:
-        repo: Git repository object
-        recursive: Recursively sync nested submodules
-
-    Returns:
-        Success or error message
-    """
-    try:
-        args = ["sync"]
-        if recursive:
-            args.append("--recursive")
-        repo.git.submodule(*args)
-        recursive_info = " (recursive)" if recursive else ""
-        return f"✅ Submodule URLs synced from .gitmodules{recursive_info}"
-    except GitCommandError as e:
-        return f"❌ Submodule sync failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Submodule sync error: {str(e)}"
-
-
-# ---------------------------------------------------------------------------
-# Git config operations
-# ---------------------------------------------------------------------------
-
-_CONFIG_KEY_RE = re.compile(r"^[a-zA-Z0-9]([a-zA-Z0-9._/\\-]*[a-zA-Z0-9])?$")
-_CONFIG_SCOPE_MAP = {
-    "local": {},
-    "global": {"_global": True},
-    "system": {"system": True},
-}
-
-
-def _build_config_scope_kwargs(scope: str | None) -> dict:
-    """Return GitPython keyword arguments for the requested config scope."""
-    if scope is None or scope == "local":
-        return {}
-    if scope not in _CONFIG_SCOPE_MAP:
-        raise ValueError(
-            f"Invalid scope '{scope}'. Valid values: local, global, system"
-        )
-    return _CONFIG_SCOPE_MAP[scope]
-
-
-def _validate_config_key(key: str) -> None:
-    """Raise ValueError if key contains characters that could cause injection."""
-    if not _CONFIG_KEY_RE.match(key):
-        raise ValueError(
-            f"Invalid config key '{key}'. Keys must contain only alphanumeric characters, "
-            "dots, dashes, and underscores (e.g., 'user.name', 'core.filemode')"
-        )
-
-
-def _validate_config_file(repo: Repo, file: str) -> None:
-    """Raise ValueError if file path attempts to escape the repository directory."""
-    # Reject any path that contains '..' components
-    file_path = Path(file)
-    if ".." in file_path.parts:
-        raise ValueError(
-            f"Invalid file path '{file}': path must not contain '..' components"
-        )
-    # Reject absolute paths that are outside the repo
-    if file_path.is_absolute():
-        try:
-            file_path.relative_to(Path(repo.working_dir))
-        except ValueError:
-            raise ValueError(
-                f"Invalid file path '{file}': absolute path must be within the repository"
-            )
-    # Resolve to catch symlink escapes (belt-and-suspenders)
-    repo_root = Path(repo.working_dir).resolve()
-    resolved = (repo_root / file_path).resolve()
-    if not str(resolved).startswith(str(repo_root)):
-        raise ValueError(f"Config file path escapes repository: {file}")
-
-
-def git_config_get(
-    repo: Repo,
-    key: str,
-    file: str | None = None,
-    scope: str | None = None,
-) -> str:
-    """Read a git config value.
-
-    Args:
-        repo: Git repository object
-        key: Config key to read (e.g., 'user.name', 'core.filemode')
-        file: Path to a specific config file (e.g., '.gitmodules')
-        scope: Config scope: 'local' (default), 'global', or 'system'
-
-    Returns:
-        The config value or an error message
-    """
-    try:
-        _validate_config_key(key)
-        scope_kwargs = _build_config_scope_kwargs(scope)
-        if file:
-            _validate_config_file(repo, file)
-            value = repo.git.config(key, f=file, **scope_kwargs)
-        else:
-            value = repo.git.config(key, **scope_kwargs)
-        return f"{key} = {value}"
-    except ValueError as e:
-        return f"❌ {e}"
-    except GitCommandError as e:
-        return f"❌ Config get failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Config get error: {str(e)}"
-
-
-def git_config_set(
-    repo: Repo,
-    key: str,
-    value: str,
-    file: str | None = None,
-    scope: str | None = None,
-) -> str:
-    """Set a git config value.
-
-    Args:
-        repo: Git repository object
-        key: Config key to set (e.g., 'user.name', 'core.filemode')
-        value: Value to set
-        file: Path to a specific config file (e.g., '.gitmodules')
-        scope: Config scope: 'local' (default), 'global', or 'system'
-
-    Returns:
-        Success or error message
-    """
-    try:
-        _validate_config_key(key)
-        scope_kwargs = _build_config_scope_kwargs(scope)
-        if file:
-            _validate_config_file(repo, file)
-            repo.git.config(key, value, f=file, **scope_kwargs)
-        else:
-            repo.git.config(key, value, **scope_kwargs)
-        scope_info = f" [{scope}]" if scope else ""
-        file_info = f" in {file}" if file else ""
-        return f"✅ Config set{scope_info}{file_info}: {key} = {value}"
-    except ValueError as e:
-        return f"❌ {e}"
-    except GitCommandError as e:
-        return f"❌ Config set failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Config set error: {str(e)}"
-
-
-def git_config_list(
-    repo: Repo,
-    file: str | None = None,
-    scope: str | None = None,
-) -> str:
-    """List all git config entries.
-
-    Args:
-        repo: Git repository object
-        file: Path to a specific config file to list (e.g., '.gitmodules')
-        scope: Config scope: 'local' (default), 'global', or 'system'
-
-    Returns:
-        Formatted list of config entries or an error message
-    """
-    try:
-        scope_kwargs = _build_config_scope_kwargs(scope)
-        if file:
-            _validate_config_file(repo, file)
-            output = repo.git.config("--list", f=file, **scope_kwargs)
-        else:
-            output = repo.git.config("--list", **scope_kwargs)
-        if not output.strip():
-            return "No config entries found"
-        source = file if file else (scope or "local")
-        return f"Git config [{source}]:\n{output}"
-    except ValueError as e:
-        return f"❌ {e}"
-    except GitCommandError as e:
-        return f"❌ Config list failed: {str(e)}"
-    except Exception as e:
-        return f"❌ Config list error: {str(e)}"
